move uppercase plugin to default
