--- plugins/all-uppercase/manifest.json
@@ -1,21 +1,0 @@
-{
-  "id": "uppercase",
-  "name": "Uppercase check",
-  "description": "Passes when the text is fully uppercase.",
-  "credentials": [],
-  "functions": [
-    {
-      "name": "All Uppercase Check",
-      "id": "alluppercase",
-      "type": "guardrail",
-      "supportedHooks": ["beforeRequest", "afterRequest"],
-      "description": [
-        {
-          "type": "subHeading",
-          "text": "Checks if all characters are uppercase."
-        }
-      ],
-      "parameters": []
-    }
-  ]
-}
